Remove unused files + re-indent
