--- RecorderTests/Mac/MacSpec.swift
@@ -1,16 +1,0 @@
-@testable import Recorder
-import Quick
-import Nimble
-
-class MacSpec: QuickSpec {
-
-  override func spec() {
-    describe("Mac") {
-      describe("test") {
-        it("passes") {
-          expect(true).to(beTrue())
-        }
-      }
-    }
-  }
-}
