--- RecorderTests/Shared/SharedSpec.swift
@@ -1,16 +1,0 @@
-@testable import Recorder
-import Quick
-import Nimble
-
-class SharedSpec: QuickSpec {
-
-  override func spec() {
-    describe("Shared") {
-      describe("test") {
-        it("passes") {
-          expect(true).to(beTrue())
-        }
-      }
-    }
-  }
-}
