--- RecorderTests/iOS/iOSSpec.swift
@@ -1,16 +1,0 @@
-@testable import Recorder
-import Quick
-import Nimble
-
-class iOSSpec: QuickSpec {
-
-  override func spec() {
-    describe("iOS") {
-      describe("test") {
-        it("passes") {
-          expect(true).to(beTrue())
-        }
-      }
-    }
-  }
-}
