--- Sources/Mac/Mac.swift
@@ -1,1 +1,0 @@
-import Cocoa
