--- Sources/iOS/iOS.swift
@@ -1,1 +1,0 @@
-import UIKit
